requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
confluent-kafka==2.0.2
fastapi==0.92.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
influxdb-client==1.36.1
paho-mqtt==1.6.1
pydantic==1.10.5
python-dateutil==2.8.2
reactivex==4.0.4
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
zipp==3.15.0
